heavy refactor of generation path
